Conflicts:
	IntelAnalytics/intel_analytics/table/bigdataframe.py
	IntelAnalytics/intel_analytics/tests/test_hbase_table.py

+++ scripts/rpm/SPECS/python-intelanalytics-dependencies.spec
@@ -1,6 +1,6 @@
-Summary: Intel Graph Analytics System development libraries
+Summary: Intel Graph Analytics System development libraries. Build number: %{?BUILD_NUMBER}. Time %{?TIMESTAMP}.
 
-Version: 0.5.0
+Version: 0.8.0
 
 License: Apache
 
@@ -8,7 +8,7 @@ Group: Development
 
 Name: python-intelanalytics-dependencies
 
-Requires: java >= 1.7, blas,  bzip2-devel,  dos2unix,  freetype-devel,  gcc,  gtk2-devel,  libffi-devel,  libpng-devel,  ncurses-devel,  openssl-devel,  pygtk2-devel,  python-devel,  readline-devel,  sqlite-devel,  tk-devel,  tkinter, atlas, atlas-devel, blas-devel, freetype, freetype-devel, gcc-c++, lapack, lapack-devel, libpng-devel, python-devel, python-setuptools, yum-utils, zlib-devel, boost-devel, intelanalytics
+Requires: java >= 1.7, blas,  bzip2-devel,  dos2unix,  freetype-devel,  gcc,  gtk2-devel,  libffi-devel,  libpng-devel,  ncurses-devel,  openssl-devel,  pygtk2-devel,  python-devel,  readline-devel,  sqlite-devel,  tk-devel,  tkinter, atlas, atlas-devel, blas-devel, freetype, freetype-devel, gcc-c++, lapack, lapack-devel, libpng-devel, python-devel, python-setuptools, yum-utils, zlib-devel, boost-devel, patch, perl-libwww-perl, intelanalytics
 
 Prefix: /usr
 
@@ -21,10 +21,23 @@ URL: <TODO>
 Buildroot: /tmp/intelanalyticsrpm
 
 %description
+Install IPython and Intel Graph System Python dependencies. Build number: %{?BUILD_NUMBER}. Time %{?TIMESTAMP}.
 
-The Intel Graph System Python libraries
+%define TIMESTAMP %(echo $TIMESTAMP)
 
 %prep
+
+url="http://www.w3.org"
+timeout=20
+result=`HEAD -d -t $timeout $url`
+ 
+if [ "$result" = "200 OK" ]; then
+    echo "Web connectivity present"
+else
+    echo "Could not connect to download dependencies. Please check your proxy settings"
+    echo "Aborting Install"
+    exit 1
+fi
 
 %setup -q
 
@@ -46,8 +59,10 @@ tar xvf $RPM_BUILD_ROOT/usr/lib/IntelAnalytics/template_overrides.tar.gz -C /usr
 ln -sf /usr/lib/IntelAnalytics/virtpy/bin/activate %{_bindir}/virtpy
 
 %postun
-rm -rf /usr/lib/IntelAnalytics/virtpy #remove vitual python 
-rm %{_bindir}/virtpy
+if [ "$1" = "0" ]; then # $1 is set to 0 for rpm uninstall and 1 for update
+  rm -rf /usr/lib/IntelAnalytics/virtpy #remove vitual python 
+  rm %{_bindir}/virtpy
+fi
 
 %files
 %{_exec_prefix}/lib/IntelAnalytics
